+++ Source/YaOpt/Patches/Compatibility/RomanceOnTheRim/RomanceOnTheRim_RomanticInteractionUtility_MakeLoverBubble.cs
@@ -1,0 +1,28 @@
+using HarmonyLib;
+using System.Reflection;
+
+namespace YaOpt.Patches.Compatibility.RomanceOnTheRim
+{
+	[HarmonyPatch]
+	[HarmonyPriority(Priority.High)]
+	internal static class RomanceOnTheRim_RomanticInteractionUtility_MakeLoverBubble
+	{
+		static MethodBase TargetMethod()
+		{
+			return AccessTools.Method(
+				AccessTools.TypeByName("RomanceOnTheRim.RomanticInteractionUtility"),
+				"MakeLoverBubble");
+		}
+
+		static bool Prepare()
+		{
+			return YaOptGlobal.NeedThreadSafe &&
+				   YaOptGlobal.HasMod("telardo.romanceontherim");
+		}
+
+		static bool Prefix()
+		{
+			return YaOptGlobal.IsInMainThread;
+		}
+	}
+}