+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Compatibility: Romance On The Rim (Imperfect; when Parallel Thought Update is enabled, the thoughts "missing lover" and "worrying about lover's health" do not display thought bubbles)
+
 ## 1.1.2  (2026-06-21)
 
 * Compatibility: Performance Optimizer - Faster GetComp methods replacement option.